Refactor ReproducibleDirectoryWalker to use the same visitor as DefaultDirectoryWalker
