+++ subprojects/file-collections/src/main/java/org/gradle/api/internal/file/collections/ReproducibleDirectoryWalker.java
@@ -16,21 +16,22 @@
 
 package org.gradle.api.internal.file.collections;
 
-import org.apache.commons.io.comparator.PathFileComparator;
 import org.gradle.api.GradleException;
 import org.gradle.api.file.FileTreeElement;
-import org.gradle.api.file.FileVisitDetails;
 import org.gradle.api.file.FileVisitor;
 import org.gradle.api.file.RelativePath;
-import org.gradle.api.internal.file.DefaultFileVisitDetails;
 import org.gradle.api.specs.Spec;
 import org.gradle.internal.nativeintegration.filesystem.FileSystem;
 
-import java.io.File;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import java.io.IOException;
+import java.nio.file.FileVisitResult;
+import java.nio.file.Files;
+import java.nio.file.LinkOption;
+import java.nio.file.Path;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.util.Comparator;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.stream.Stream;
 
 public class ReproducibleDirectoryWalker implements DirectoryWalker {
     private final FileSystem fileSystem;
@@ -39,49 +40,50 @@ public class ReproducibleDirectoryWalker implements DirectoryWalker {
         this.fileSystem = fileSystem;
     }
 
-    protected File[] getChildren(File file) {
-        File[] children = file.listFiles();
-        if (children != null) {
-            Arrays.sort(children, PathFileComparator.PATH_COMPARATOR);
-        }
-        return children;
-    }
-
     @Override
-    public void walkDir(File file, RelativePath path, FileVisitor visitor, Spec<? super FileTreeElement> spec, AtomicBoolean stopFlag, boolean postfix) {
-        File[] children = getChildren(file);
-        if (children == null) {
-            if (file.isDirectory() && !file.canRead()) {
-                throw new GradleException(String.format("Could not list contents of directory '%s' as it is not readable.", file));
-            }
-            // else, might be a link which points to nothing, or has been removed while we're visiting, or ...
-            throw new GradleException(String.format("Could not list contents of '%s'.", file));
+    public void walkDir(Path rootDir, RelativePath rootPath, FileVisitor visitor, Spec<? super FileTreeElement> spec, AtomicBoolean stopFlag, boolean postfix) {
+        try {
+            PathVisitor pathVisitor = new PathVisitor(spec, postfix, visitor, stopFlag, rootPath, fileSystem);
+            visit(rootDir, pathVisitor);
+        } catch (IOException e) {
+            throw new GradleException(String.format("Could not list contents of directory '%s'.", rootDir), e);
         }
-        List<FileVisitDetails> dirs = new ArrayList<FileVisitDetails>();
-        for (int i = 0; !stopFlag.get() && i < children.length; i++) {
-            File child = children[i];
-            boolean isFile = child.isFile();
-            RelativePath childPath = path.append(isFile, child.getName());
-            FileVisitDetails details = new DefaultFileVisitDetails(child, childPath, stopFlag, fileSystem, fileSystem);
-            if (DirectoryFileTree.isAllowed(details, spec)) {
-                if (isFile) {
-                    visitor.visitFile(details);
-                } else {
-                    dirs.add(details);
-                }
+    }
+
+    private static FileVisitResult visit(Path path, PathVisitor pathVisitor) throws IOException {
+        BasicFileAttributes attrs;
+        try {
+            attrs = Files.readAttributes(path, BasicFileAttributes.class);
+        } catch (IOException e) {
+            try {
+                attrs = Files.readAttributes(path, BasicFileAttributes.class, LinkOption.NOFOLLOW_LINKS);
+            } catch (IOException next) {
+                return pathVisitor.visitFileFailed(path, next);
             }
         }
 
-        // now handle dirs
-        for (int i = 0; !stopFlag.get() && i < dirs.size(); i++) {
-            FileVisitDetails dir = dirs.get(i);
-            if (postfix) {
-                walkDir(dir.getFile(), dir.getRelativePath(), visitor, spec, stopFlag, postfix);
-                visitor.visitDir(dir);
-            } else {
-                visitor.visitDir(dir);
-                walkDir(dir.getFile(), dir.getRelativePath(), visitor, spec, stopFlag, postfix);
+        if (attrs.isDirectory()) {
+            FileVisitResult fvr = pathVisitor.preVisitDirectory(path, attrs);
+            if (fvr == FileVisitResult.SKIP_SUBTREE || fvr == FileVisitResult.TERMINATE) {
+                return fvr;
             }
+            IOException exception = null;
+            try (Stream<Path> fileStream = Files.list(path)) {
+                Iterable<Path> files = () -> fileStream.sorted(FILES_FIRST).iterator();
+                for (Path child : files) {
+                    FileVisitResult childResult = visit(child, pathVisitor);
+                    if (childResult == FileVisitResult.TERMINATE) {
+                        return childResult;
+                    }
+                }
+            } catch (IOException e) {
+                exception = e;
+            }
+            return pathVisitor.postVisitDirectory(path, exception);
+        } else {
+            return pathVisitor.visitFile(path, attrs);
         }
     }
+
+    private final static Comparator<Path> FILES_FIRST = Comparator.<Path, Boolean>comparing(x -> x.toFile().isDirectory()).thenComparing(Path::toString);
 }
